migrate Primer::LinkComponent to Primer::Beta::LinkComponent (#1532)
