+++ .changeset/happy-yaks-smoke.md
@@ -1,0 +1,5 @@
+---
+'@primer/view-components': patch
+---
+
+migrate Primer::LinkComponent to Primer::Beta::Link and mark the old one as deprecated
