+++ lib/primer/deprecations.rb
@@ -6,6 +6,7 @@ module Primer
     # If there is no alternative to suggest, set the value to nil
     DEPRECATED_COMPONENTS = {
       "Primer::LabelComponent" => "Primer::Beta::Label",
+      "Primer::LinkComponent" => "Primer::Beta::Link",
       "Primer::ImageCrop" => "Primer::Alpha::ImageCrop",
       "Primer::Image" => "Primer::Alpha::Image",
       "Primer::Alpha::AutoComplete" => "Primer::Beta::AutoComplete",
