+++ previews/primer/alpha/tooltip_preview.rb
@@ -75,9 +75,9 @@ module Primer
         end
       end
 
-      # @label Tooltip with Primer::LinkComponent
+      # @label Tooltip with Primer::Beta::Link
       def tooltip_with_link(type: :description, direction: :s, tooltip_text: "Tooltip text")
-        render(Primer::LinkComponent.new(href: "#link-with-tooltip", id: "link-with-tooltip")) do |c|
+        render(Primer::Beta::Link.new(href: "#link-with-tooltip", id: "link-with-tooltip")) do |c|
           c.tooltip(text: tooltip_text, type: type, direction: direction)
           "Button"
         end
@@ -85,7 +85,7 @@ module Primer
 
       # @label Tooltip with Primer::IconButton
       def tooltip_with_icon_button(direction: :s, tooltip_text: "Tooltip text")
-        render(Primer::IconButton.new(icon: :search, "aria-label": tooltip_text, tooltip_direction: direction))
+        render(Primer::Beta::IconButton.new(icon: :search, "aria-label": tooltip_text, tooltip_direction: direction))
       end
       # @!endgroup
     end
