+++ static/audited_at.json
@@ -48,6 +48,7 @@
   "Primer::Beta::Heading": "",
   "Primer::Beta::IconButton": "",
   "Primer::Beta::Label": "",
+  "Primer::Beta::Link": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
   "Primer::Beta::Truncate::TruncateText": "",
