+++ static/constants.json
@@ -636,6 +636,19 @@
       "none"
     ]
   },
+  "Primer::Beta::Link": {
+    "DEFAULT_SCHEME": "default",
+    "DEFAULT_TAG": "a",
+    "SCHEME_MAPPINGS": {
+      "default": "",
+      "primary": "Link--primary",
+      "secondary": "Link--secondary"
+    },
+    "TAG_OPTIONS": [
+      "a",
+      "span"
+    ]
+  },
   "Primer::Beta::Text": {
     "DEFAULT_TAG": "span"
   },
@@ -782,17 +795,6 @@
     "MAX_COL": 12
   },
   "Primer::LinkComponent": {
-    "DEFAULT_SCHEME": "default",
-    "DEFAULT_TAG": "a",
-    "SCHEME_MAPPINGS": {
-      "default": "",
-      "primary": "Link--primary",
-      "secondary": "Link--secondary"
-    },
-    "TAG_OPTIONS": [
-      "a",
-      "span"
-    ]
   },
   "Primer::LocalTime": {
     "DEFAULT_DIGIT_TYPE": "numeric",
