+++ static/statuses.json
@@ -48,6 +48,7 @@
   "Primer::Beta::Heading": "beta",
   "Primer::Beta::IconButton": "beta",
   "Primer::Beta::Label": "beta",
+  "Primer::Beta::Link": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -69,7 +70,7 @@
   "Primer::ImageCrop": "deprecated",
   "Primer::LabelComponent": "deprecated",
   "Primer::LayoutComponent": "alpha",
-  "Primer::LinkComponent": "beta",
+  "Primer::LinkComponent": "deprecated",
   "Primer::LocalTime": "alpha",
   "Primer::Markdown": "beta",
   "Primer::MenuComponent": "alpha",
